requirements.txt
Flask==2.3.3
pandas
scikit-surprise
requests
